[Bronze II] Title: 수학은 비대면강의입니다, Time: 40 ms, Memory: 31256 KB -BaekjoonHub

+++ 백준/Bronze/19532. 수학은 비대면강의입니다/수학은 비대면강의입니다.py
@@ -1,4 +1,4 @@
 a,b,c,d,e,f = map(int,input().split())
 y = (f-c*d/a)/(e-b*d/a) if a else c/b
 x = (c-b*y)/a if a else (f-c*e/b)/d
-print(int(round(x)), int(round(y)))
+print(round(x), round(y))
